# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cero Colas — Medicion de tiempo de espera",
    "description": "\nAPI del sistema de medicion automatizada del tiempo de espera, en el marco del\nPlan Cero Colas (RM N.° 811-2018/MINSA).\n\n**Como funciona.** Los agentes de vision instalados en cada hospital detectan\ncruces de personas en lineas virtuales y los envian a `/v1/cruces:lote`. El\nservidor reconstruye las curvas de flujo acumulado y deriva la ocupacion y el\ntiempo de espera.\n\n**Que no hace.** No identifica personas, no almacena imagenes y no realiza\nreconocimiento facial ni re-identificacion. Un cruce solo registra que alguien\natraveso una linea en un sentido y en un instante.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/cruces:lote": {
      "post": {
        "tags": [
          "ingesta"
        ],
        "summary": "Ingesta por lotes de cruces detectados",
        "operationId": "ingestar_cruces_v1_cruces_lote_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoteCruces"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaLote"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/aforos": {
      "post": {
        "tags": [
          "ingesta"
        ],
        "summary": "Conteo directo de ocupacion",
        "operationId": "registrar_aforo_v1_aforos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AforoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Registrar Aforo V1 Aforos Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AforoEntrada": {
        "properties": {
          "sala_id": {
            "type": "string",
            "title": "Sala Id"
          },
          "ocurrido_en": {
            "type": "string",
            "format": "date-time",
            "title": "Ocurrido En"
          },
          "conteo": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Conteo"
          },
          "metodo": {
            "type": "string",
            "title": "Metodo",
            "default": "VISION"
          },
          "confianza": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confianza"
          }
        },
        "type": "object",
        "required": [
          "sala_id",
          "ocurrido_en",
          "conteo"
        ],
        "title": "AforoEntrada"
      },
...
      "CruceEntrada": {
        "properties": {
          "linea": {
            "type": "string",
            "title": "Linea",
            "description": "Nombre de la linea virtual, p. ej. ENTRADA_SALA"
          },
          "direccion": {
            "type": "string",
            "enum": [
              "IN",
              "OUT"
            ],
            "title": "Direccion"
          },
          "ocurrido_en": {
            "type": "string",
            "format": "date-time",
            "title": "Ocurrido En",
            "description": "ISO 8601 con desfase horario explicito"
          },
          "confianza": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confianza"
          },
          "track_id_local": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Id Local",
            "description": "Identificador de seguimiento local del agente. No identifica a ninguna persona: se reinicia con cada arranque y solo sirve para depurar conteos duplicados."
          }
        },
        "type": "object",
        "required": [
          "linea",
          "direccion",
          "ocurrido_en"
        ],
        "title": "CruceEntrada",
        "description": "Un cruce reportado por el agente de vision."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoteCruces": {
        "properties": {
          "sala_id": {
            "type": "string",
            "title": "Sala Id",
            "description": "Codigo de la sala, p. ej. LOR-CE-01"
          },
          "agente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agente",
            "description": "Hostname del agente emisor"
          },
          "cruces": {
            "items": {
              "$ref": "#/components/schemas/CruceEntrada"
            },
            "type": "array",
            "maxItems": 5000,
            "minItems": 1,
            "title": "Cruces"
          }
        },
        "type": "object",
        "required": [
          "sala_id",
          "cruces"
        ],
        "title": "LoteCruces"
      },
...
      "RechazoCruce": {
        "properties": {
          "indice": {
            "type": "integer",
            "title": "Indice"
          },
          "motivo": {
            "type": "string",
            "title": "Motivo"
          }
        },
        "type": "object",
        "required": [
          "indice",
          "motivo"
        ],
        "title": "RechazoCruce"
      },
...
      "RespuestaLote": {
        "properties": {
          "lote_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lote Id"
          },
          "recibidos": {
            "type": "integer",
            "title": "Recibidos"
          },
          "aceptados": {
            "type": "integer",
            "title": "Aceptados"
          },
          "rechazados": {
            "type": "integer",
            "title": "Rechazados"
          },
          "duplicado": {
            "type": "boolean",
            "title": "Duplicado",
            "description": "True si la clave de idempotencia ya habia sido procesada; en ese caso no se insertaron cruces nuevos.",
            "default": false
          },
          "detalles": {
            "items": {
              "$ref": "#/components/schemas/RechazoCruce"
            },
            "type": "array",
            "title": "Detalles",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "lote_id",
          "recibidos",
          "aceptados",
          "rechazados"
        ],
        "title": "RespuestaLote"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/v1/auth/token"
          }
        }
      }
    }
  }
}